Move WP notices component to panels directory.
